requirements.txt
Django==5.1.6
djangorestframework==3.15.2
django-cors-headers==4.7.0
django-extensions==3.2.3
requests==2.32.3
